+++ CHANGELOG.md
@@ -11,6 +11,11 @@ This Changelog tracks changes to this project. The notes below include a summary
 
 ## Version `0.0.0`
 
+### 11 Jun 2023
+
+- `changed` sqlalchemy declarative to 2.0 standard
+- `added` some basic test tables for game things
+
 ### 10 Jun 2023
 
 - `added` postgres database with declarative sqlalchemy